requirements.txt
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
nltk
pandas
numpy
beautifulsoup4
pydantic
langchain-openai
langchain_core
python-dotenv